revert plugin sdk request separation
